Separate test renderer host config
